Cs upstream 5.7.23 merge

+++ mysql-test/mysql-test-run.pl
@@ -3730,6 +3730,12 @@ sub have_wsrep() {
   return defined $wsrep_on
 }
 
+sub wsrep_is_bootstrap_server($) {
+  my $mysqld= shift;
+  return $mysqld->if_exist('wsrep_cluster_address') &&
+    ($mysqld->value('wsrep_cluster_address') eq "gcomm://" ||
+     $mysqld->value('wsrep_cluster_address') eq "'gcomm://'");
+}
 
 sub check_wsrep_support() {
   if (have_wsrep())
@@ -4302,13 +4308,8 @@ sub wait_wsrep_ready($$) {
 
   for (my $loop= 1; $loop <= $loops; $loop++)
   {
-    if (run_query_output($mysqld, $query, $outfile) != 0)
-    {
-      $tinfo->{logfile}= "WSREP error while trying to determine node state";
-      return 0;
-    }
-
-    if (mtr_grab_file($outfile) =~ /^ON/)
+    if (run_query_output($mysqld, $query, $outfile) == 0 &&
+        mtr_grab_file($outfile) =~ /^ON/)
     {
       unlink($outfile);
       return 1;
@@ -6649,6 +6650,22 @@ sub start_servers($) {
       # Save this test case information, so next can examine it
       $mysqld->{'started_tinfo'}= $tinfo;
 
+      # Wait until server's uuid is generated. This avoids that master and
+      # slave generate the same UUID sporadically.
+      sleep_until_file_created("$datadir/auto.cnf", $opt_start_timeout,
+                               $mysqld->{'proc'});
+
+      # If wsrep is on, we need to wait until the first
+      # server starts and bootstraps the cluster before
+      # starting other servers.
+      if (have_wsrep() && wsrep_is_bootstrap_server($mysqld))
+      {
+        mtr_verbose("WSREP waiting for first server to bootstrap cluster");
+        if (!wait_wsrep_ready($tinfo, $mysqld))
+        {
+          return 1;
+        }
+      }
     }
 
   }
